Moves occa command building in include and src
