+++ include/occa/bin/occa.hpp
@@ -1,0 +1,12 @@
+#ifndef OCCA_BIN_OCCA_HEADER
+#define OCCA_BIN_OCCA_HEADER
+
+#include <occa/tools/cli.hpp>
+
+namespace occa {
+  namespace bin {
+    cli::command buildOccaCommand();
+  }
+}
+
+#endif
